fake genesis with stable fake keys
